design: -Header, Footer 컴포넌트 ui 수정 -NaverMap 적용 Rename: Layout폴더 common하위로 이동
