+++ src/Router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 
 import HomePage from "@/pages/HomePage"
-import Layout from "@/components/Layout";
+import Layout from "@/components/common/Layout";
 
 const Router = () => {
   return(
