+++ src/components/Home/Home.style.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Main = styled.main`
+    div {
+        width: 100%;
+        height: calc(100vh - 136px);
+    }
+    
+`
