+++ src/components/common/Layout/Footer/index.tsx
@@ -1,7 +1,7 @@
-import {Container, Nav, NavItem} from "@/components/Layout/Footer/Footer.style.ts";
+import {Container, Nav, NavItem} from "@/components/common/Layout/Footer/Footer.style.ts";
 import TaxiIcon from "@/assets/footer/taxi-icon.svg?react";
-import ChattingIcon from  "@/assets/footer/chatting-icon.svg?react";
-import  CreateButtonIcon from  "@/assets/footer/create-button-icon.svg?react";
+import ChattingIcon from "@/assets/footer/chatting-icon.svg?react";
+import  CreateButtonIcon from "@/assets/footer/create-button-icon.svg?react";
 import HistoryIcon from "@/assets/footer/history-icon.svg?react";
 import UserIcon from "@/assets/footer/user-icon.svg?react";
