+++ src/components/common/Layout/Header/Header.style.ts
@@ -1,10 +1,9 @@
 import styled from "styled-components";
 
 export const Container = styled.header`
-    width: 100%;
     height: 36px;;
     
-    padding: 20px;
+    padding: 10px 20px;
     z-index: 999;
     
     display: flex;
