+++ src/components/common/Layout/Layout.style.ts
@@ -3,7 +3,7 @@ import styled from "styled-components";
 const Container = styled.div`
     max-width: 500px;
     width: 100%;
-    min-height: 100vh;
+    height: 100vh;
     
     display: flex;
     justify-content: center;
